infra: Make docker folder read only, remove unused tests, update repo name, remove incorrect documentation.
