+++ .gitignore
@@ -8,3 +8,4 @@ __pycache__/
 **/*.pyc
 **.pyc
 **/*.tar.gz
+.coverage
