+++ NOTICE
@@ -1,2 +1,2 @@
-sagemaker-pytorch-serving-container
+sagemaker-pytorch-inference-toolkit
 Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
