+++ artifacts/config.properties
@@ -1,0 +1,26 @@
+vmargs=-Xmx128m -XX:-UseLargePages -XX:+UseG1GC -XX:MaxMetaspaceSize=32M -XX:MaxDirectMemorySize=10m -XX:+ExitOnOutOfMemoryError
+model_store=/opt/ml/model
+load_models=ALL
+inference_address=http://0.0.0.0:8080
+management_address=http://0.0.0.0:8081
+# management_address=unix:/tmp/management.sock
+# number_of_netty_threads=0
+# netty_client_threads=0
+# default_response_timeout=120
+# default_workers_per_model=0
+# job_queue_size=100
+# async_logging=false
+# number_of_gpu=1
+# cors_allowed_origin
+# cors_allowed_methods
+# cors_allowed_headers
+# keystore=src/test/resources/keystore.p12
+# keystore_pass=changeit
+# keystore_type=PKCS12
+# private_key_file=src/test/resources/key.pem
+# certificate_file=src/test/resources/certs.pem
+# max_response_size=6553500
+# max_request_size=6553500
+# blacklist_env_vars=
+# decode_input_request=false
+# enable_envvars_config=false
