--- test-toolkit/docker/1.5.0/Dockerfile.dlc.gpu
@@ -1,6 +1,0 @@
-ARG region
-FROM 763104351884.dkr.ecr.$region.amazonaws.com/pytorch-inference:1.5.0-gpu-py3
-
-COPY dist/sagemaker_pytorch_inference-*.tar.gz /sagemaker_pytorch_inference.tar.gz
-RUN pip install --upgrade --no-cache-dir /sagemaker_pytorch_inference.tar.gz && \
-    rm /sagemaker_pytorch_inference.tar.gz
