--- test-toolkit/docker/1.5.0/Dockerfile.pytorch
@@ -1,41 +1,0 @@
-FROM pytorch/pytorch:1.5-cuda10.1-cudnn7-runtime
-
-LABEL com.amazonaws.sagemaker.capabilities.accept-bind-to-port=true
-LABEL com.amazonaws.sagemaker.capabilities.multi-models=true
-
-ARG MMS_VERSION=1.0.8
-
-ENV SAGEMAKER_SERVING_MODULE sagemaker_pytorch_serving_container.serving:main
-ENV TEMP=/home/model-server/tmp
-
-RUN apt-get update \
-    && apt-get install -y --no-install-recommends \
-    libgl1-mesa-glx \
-    libglib2.0-0 \
-    libsm6 \
-    libxext6 \
-    libxrender-dev \
-    openjdk-8-jdk-headless \
-    && rm -rf /var/lib/apt/lists/*
-
-RUN conda install -c conda-forge opencv==4.0.1 \
-    && ln -s /opt/conda/bin/pip /usr/local/bin/pip3
-
-RUN pip install mxnet-model-server==$MMS_VERSION
-
-COPY dist/sagemaker_pytorch_inference-*.tar.gz /sagemaker_pytorch_inference.tar.gz
-RUN pip install --no-cache-dir /sagemaker_pytorch_inference.tar.gz && \
-    rm /sagemaker_pytorch_inference.tar.gz
-
-RUN useradd -m model-server \
-    && mkdir -p /home/model-server/tmp \
-    && chown -R model-server /home/model-server
-
-COPY docker/build_artifacts/mms-entrypoint.py /usr/local/bin/dockerd-entrypoint.py
-COPY docker/build_artifacts/config.properties /home/model-server
-
-RUN chmod +x /usr/local/bin/dockerd-entrypoint.py
-
-EXPOSE 8080 8081
-ENTRYPOINT ["python", "/usr/local/bin/dockerd-entrypoint.py"]
-CMD ["mxnet-model-server", "--start", "--mms-config", "/home/model-server/config.properties"]
