+++ tox.ini
@@ -15,7 +15,8 @@ exclude =
     __pycache__
     .tox
     test/resources/
-    test-toolkit/resources/
+    artifacts/
+    docker/
     lib/
 max-complexity = 10
 ignore =
